Fix compiler warnings (#6113)
* Fix: Refactor loop indices to use size_t for better type safety

* Refactor: Remove unused variables and clean up code in multiple files

* Fix: Change color_index to size_t for improved type safety in create_surface method

* Fix: Correct logical grouping in shader enablement condition across multiple renderer files

* Refactor: Remove unused compileShader function from Renderer and ShaderUtils

* Fix: Correct order of member variable declarations in FBO class

* Fix: Adjust diagnostic pragmas to get rid of dangling pointer warnings in StackCheck

* Refactor: Change currentlySelectedObject from int to std::optional<size_t> for improved type safety

+++ src/io/export_3mf_v1.cc
@@ -240,7 +240,7 @@ bool append_polyset(const std::shared_ptr<const PolySet>& ps, ExportContext& ctx
     }
 
     ctx.materialids.reserve(sorted_ps->colors.size());
-    for (int i = 0; i < sorted_ps->colors.size(); i++) {
+    for (size_t i = 0; i < sorted_ps->colors.size(); i++) {
       ctx.materialids.push_back(materialFunc(materials + i, sorted_ps->colors[i]));
     }
   }
@@ -251,7 +251,7 @@ bool append_polyset(const std::shared_ptr<const PolySet>& ps, ExportContext& ctx
     return false;
   }
 
-  for (int i = 0; i < sorted_ps->color_indices.size(); ++i) {
+  for (size_t i = 0; i < sorted_ps->color_indices.size(); ++i) {
     const int32_t idx = sorted_ps->color_indices[i];
     if (!handle_triangle_color(propertyhandler, sorted_ps, i, idx, ctx)) {
       return false;
